tests/wpt/web-platform-tests/svg/scripted/tspan-attrs-xyrotate-have-length.svg

<?xml version="1.0" standalone="no"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml">
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/single-page.html#text-TSpanNotes"/>
    <h:meta name="assert" content="dx and dy attributes on text elements are lists that support length()"/>
</metadata>
  <text id="text" font-family="Verdana" font-size="55" fill="blue"
    y="150"
    x="120"
    >Her socks were <tspan id="tsp"
    x="280 325" y="150 155 160 165 170" rotate="-30,0,30"
    >stained</tspan> with blood.</text>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <script><![CDATA[
    test(function() {
      var tspan = document.getElementById('tsp');
      assert_equals(tspan.x.baseVal.length, 2);
      assert_equals(tspan.y.baseVal.length, 5);
      assert_equals(tspan.rotate.baseVal.length, 3);
    });
  ]]></script>
</svg>
